feat: create reusable delete function and refactor file structure for improved maintainability
