--- lib/network/queries.ts
@@ -1,7 +1,0 @@
-// lib/queries.ts
-import axios from "axios";
-
-export const fetchProducts = async () => {
-  const res = await axios.get("/api/products");
-  return res.data;
-};
